requirements.txt
click==7.1.1
dash==1.9.1
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jupyter-client==6.1.5
jupyter-core==4.6.3
jupyter-server==1.11.2
lxml==4.6.3
matplotlib==3.4.1
notebook==6.0.3
numpy==1.18.2
pandas==1.2.1
pandoc==1.0.2
pandocfilters==1.4.2
Pillow==8.2.0
plotly==4.5.4
pylint==2.12.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.11.13
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
sklearn==0.0
smart-open==5.0.0
soupsieve==2.2.1
spacy==3.0.5
spacy-legacy==3.0.2
srsly==2.4.0
tornado==6.0.4
tqdm==4.59.0
traitlets==4.3.3
tornado==6.1
typing-extensions==3.7.4
urllib3==1.25.9
xgboost==1.5.2
xlrd==1.2.0
yfinance==0.1.67
zipp==3.1.0